moved ReportsLineChart to components
